Add Create Group Screen
This screen is used to add a new group. The person who performs this action is the host of the group. To create a new group an host has to declare, for the group:
- Name
- Budget
- Theme (or gift category, such as 'book', 'cook', a color or else)
- Date of gifts exchange

The host that creates the group is automatically registered as a group member, so he is automatically joins the secret santa.